Fix i2c function not writing properly due to if else and updated other functions to use it

+++ src/sensor/retrieveSenData.cpp
@@ -11,14 +11,14 @@ RetrieveSenData::RetrieveSenData(I2cControl *i2c_controller, unsigned char I2C_A
 
 void RetrieveSenData::get_LocalWeatherData(WeatherSensor *ptr_weatherSensorData)
 {
-    i2c_controller->writeInt(I2C_ADDR, 0);
+    i2c_controller->writeByte(I2C_ADDR, 0);
     usleep(50000);
     i2c_controller->readI2c(ptr_weatherSensorData->weatherSensorUnion.packet, ptr_weatherSensorData->get_temporaryStructSize());
 }
 
 void RetrieveSenData::get_RemoteWeatherSenData(WeatherSensor *ptr_weatherSensorData)
 {
-    i2c_controller->writeInt(I2C_ADDR, 1);
+    i2c_controller->writeByte(I2C_ADDR, 1);
     usleep(50000);
     i2c_controller->readI2c(ptr_weatherSensorData->weatherSensorUnion.packet, ptr_weatherSensorData->get_temporaryStructSize());
 }
